oll-doc: Allow saving to HTML files
HtmlFile.save()
Snippet.saveToHtml()
Snippets.saveToHtml()
MainWindow.saveSnippetsToHtml()

write out the content of one or all snippets to corresponding files
in the /doc folder.

Currently there is no user interface for this, but I've tested that
it works correctly. To be useful this needs an alternative HTML output
for the HtmlFile class because then the navigation has to be done
*in* the page and not outside of it in the application.