oll-doc: Add initial \markup parsing
Try to convert markup commands like \bold or the custom \ollCommand to
marked up HTML. This is necessary for the custom header fields that are
entered as \markup. With this LilyPondMarkup(object) class it is
possible to use markup commands that show correctly both in the compiled
LilyPond files and the HTML documentation.

Currently implemented:
- \ollCommand
- \italic
- \bold
- \typewriter

Items are correctly handled when the keyword is either followed by a
single token or by an expression surrounded by curly braces. Nesting
is not supported so far.