devMode: Initial implementation
"devMode" is a state indicating that a score is still under
development. oll-files (or individual functions) can use this
state to visually highlight specific operations, presumably
through colors.

When including oll-base a variable 'dev-mode is initialized to #f
so any function can base a conditional expression on the boolean
value of 'dev-mode.