scholarly: Add stub for raw Scheme export
Starts #138

As of the discussion in
http://lists.gnu.org/archive/html/lilypond-user/2015-11/msg00186.html
(and maybe also
http://lists.gnu.org/archive/html/lilypond-user/2015-11/msg00187.html)
we want to export annotations in a Scheme representation that can later
be used to write out a critical report withing LilyPond.

This commit provides the basic integration into the annotate architecture,
but the actual formatting of the written file is non-functional:
The string representations of the different objects can't be read in again
later.
Presumably it'll be necessary to implement a specific formatting routine
as for plaintext and latex output.